Make `FileManagerProvider` a dependency, make `AttachmentStore` use `@Dependency` macro for `FileManagerProvider`
