+++ TrekPoint/Managers/AttachmentProvider/AttachmentStore.swift
@@ -1,6 +1,7 @@
 import Foundation
 import UIKit
 import SwiftUI
+import Dependencies
 
 enum AttachmentError: Error {
     case fileNotFound
@@ -10,33 +11,13 @@ enum AttachmentError: Error {
     case fileReadFailed(Error)
 }
 
-protocol FileManagerProvider {
-    var documentDirectory: URL { get }
-    
-    func fileExists(atPath: String) -> Bool
-    func createDirectory(at url: URL, withIntermediateDirectories: Bool, attributes: [FileAttributeKey : Any]?) throws
-    func createFile(atPath path: String, contents: Data?, attributes: [FileAttributeKey : Any]?) -> Bool
-    func copyItem(at source: URL, to destination: URL) throws
-    func setAttributes(_ attributes: [FileAttributeKey : Any], ofItemAtPath path: String) throws
-    func removeItem(at url: URL) throws
-}
-
-extension FileManager: FileManagerProvider {
-    var documentDirectory: URL {
-        urls(for: .documentDirectory, in: .userDomainMask)[0]
-    }
-}
-
 class AttachmentStore: AttachmentProvider {
-    private let fileManager: FileManagerProvider
-    private let attachmentsDirectory: URL
+    @Dependency(\.fileManager) var fileManager
+    private var attachmentsDirectory: URL {
+        fileManager.documentDirectory.appendingPathComponent("Attachments", isDirectory: true)
+    }
     
-    init(fileManager: FileManagerProvider = FileManager.default) {
-        self.fileManager = fileManager
-        
-        let documentsDirectory = fileManager.documentDirectory
-        let attachmentsDirectory = documentsDirectory.appendingPathComponent("Attachments", isDirectory: true)
-        
+    init() {
         // Create directory if it doesn't exist
         if !fileManager.fileExists(atPath: attachmentsDirectory.path()) {
             do {
@@ -45,8 +26,6 @@ class AttachmentStore: AttachmentProvider {
                 fatalError("Failed to create Attachments directory: \(error)")
             }
         }
-        
-        self.attachmentsDirectory = attachmentsDirectory
     }
     
     func storeImage(_ image: UIImage) throws -> Attachment {
