+++ TrekPointTests/AttachmentStoreTests.swift
@@ -1,44 +1,7 @@
 @testable import TrekPoint
 import Testing
 import UIKit
-
-// TODO: - TestFileManager only mocks the happy path
-// I should conditionally allow errors to be thrown and test my assumptions there too
-// if I run into issues/want to expand my test suite
-
-class TestFileManager: FileManagerProvider {
-    let documentDirectory = URL.temporaryDirectory
-    var files = [String : (contents: Data, attributes: [FileAttributeKey : Any])]()
-    var removedFilePaths = [String]()
-    var copiedItems = [(source: URL, destination: URL)]()
-    
-    func fileExists(atPath path: String) -> Bool {
-        files[path] != nil
-    }
-    
-    func createDirectory(at url: URL, withIntermediateDirectories: Bool, attributes: [FileAttributeKey : Any]?) throws {}
-    
-    func createFile(atPath path: String, contents: Data?, attributes: [FileAttributeKey : Any]?) -> Bool {
-        files[path] = (contents ?? Data(), attributes ?? [:])
-        
-        return true
-    }
-    
-    func copyItem(at source: URL, to destination: URL) throws {
-        files[destination.path()] = files[source.path()]
-        copiedItems.append((source, destination))
-    }
-    
-    func setAttributes(_ attributes: [FileAttributeKey : Any], ofItemAtPath path: String) throws {
-        files[path]?.attributes = attributes
-    }
-    
-    func removeItem(at url: URL) throws {
-        files[url.path()] = nil
-        
-        removedFilePaths.append(url.path())
-    }
-}
+import Dependencies
 
 @Suite
 struct AttachmentStoreTests {
@@ -46,8 +9,10 @@ struct AttachmentStoreTests {
     let store: AttachmentStore
     
     init() {
-        self.fileManager = TestFileManager()
-        self.store = AttachmentStore(fileManager: fileManager)
+        // Initialize AttachmentStore directly so that the test
+        // value of AttachmentProvider isn't used (TestAttachmentStore)
+        self.store = AttachmentStore()
+        self.fileManager = store.fileManager as! TestFileManager
     }
     
     func createTestImage() -> UIImage {
@@ -79,9 +44,6 @@ struct AttachmentStoreTests {
     
     @Test("Store image throws when image compression fails")
     func storeImageThrowsOnCompressionFailure() throws {
-        let mockFileManager = TestFileManager()
-        let store = AttachmentStore(fileManager: mockFileManager)
-        
         // Image that can't be compressed to JPEG
         let mockImage = UIImage()
         
@@ -146,7 +108,6 @@ struct AttachmentStoreTests {
     func getUrlThrowsWhenFileDoesntExist() {
         let attachment = Attachment(type: .image)
         
-        // Act & Assert
         do {
             _ = try store.getUrl(for: attachment)
             #expect(Bool(false), "Should have thrown an error")
